requirements.txt
google-generativeai>=0.5.0
python-dotenv>=1.0.0
streamlit>=1.30.0
pypdf>=3.0.0
